requirements.txt
PyYAML
beautifulsoup4==4.6.0
requests==2.18.4